group_b1_cw3/requirements.txt
absl-py==2.0.0
aiohttp==3.7.4
alabaster==0.7.12
allensdk==2.15.2
anyjson==0.3.3
appdirs==1.4.4
apteryx==0.7.2
argon2-cffi==21.1.0
argschema==3.0.4
arrow==1.2.1
astroid==2.9.3
asttokens==2.0.5
astunparse==1.6.3
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
Babel==2.8.0
backcall==0.2.0
beniget==0.4.1
black==21.10b0
bleach==4.1.0
blinker==1.4
blosc2==2.0.0
boto3==1.17.21
botocore==1.20.112
Bottleneck==1.3.2
Brlapi==0.8.3
Brotli==1.0.9
cachetools==4.2.4
certifi==2020.6.20
chardet==3.0.4
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
command-not-found==0.3
conda-package-handling==1.7.3
cosignego===1.1d
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
cypari2==2.1.2
Cython==0.29.28
dbus-python==1.2.18
debmake==4.3.2
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
devscripts===2.22.1ubuntu1
diff-match-patch==20200713
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.1.0
docutils==0.17.1
duplicity==0.8.21
entrypoints==0.4
Farama-Notifications==0.0.4
fasteners==0.14.1
filelock==3.6.0
flatbuffers==23.5.26
fonttools==4.29.1
fpylll==0.5.2.dev0
fs==2.4.12
future==0.18.2
gast==0.5.4
Glymur==0.8.19
gmpy2==2.1.2
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.14.2
greenlet==1.1.2
grpcio==1.59.0
gssapi==1.6.12
gymnasium==0.29.0
gyp==0.1
h5py==3.9.0
hdmf==3.4.7
helpdev==0.7.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
igraph==0.9.9
imageio==2.31.4
imagesize==1.3.0
imaplib2==3.5
importlib-metadata==4.6.4
iniconfig==1.1.1
intervaltree==3.0.2
iotop==0.6
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
ipywidgets==6.0.0
isort==5.6.4
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==0.17.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
keras==2.14.0
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==0.0.0
lazy_loader==0.3
ldap3==2.8.1
libclang==16.0.6
lockfile==0.12.2
logilab-common==1.8.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
Markdown==3.5
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.4.2
matplotlib-inline==0.1.3
mccabe==0.6.1
meld==3.20.4
memory-allocator==0.1.2
mercurial==6.1.1
ml-dtypes==0.2.0
mock==4.0.3
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.6
multidict==6.0.4
mypy==0.942
mypy-extensions==0.4.3
nbclient==0.5.6
nbconvert==6.4.0
nbformat==5.1.3
ndx-events==0.2.0
nest-asyncio==1.5.4
netifaces==0.11.0
networkx==3.1
NEURON==7.6
nltk==3.7
nose==1.3.7
notebook==6.4.8
numexpr==2.8.6
numpy==1.23.5
numpydoc==1.2
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.1
pathspec==0.9.0
patsy==0.5.3
pbr==5.8.0
pdfarranger==1.8.2
pexpect==4.8.0
pickleshare==0.7.5
pikepdf==5.0.1+dfsg
Pillow==9.0.1
pipenv==11.9.0
pkgconfig==1.5.5
platformdirs==2.5.1
pluggy==0.13.0
ply==3.11
pplpy==0.8.7
primecountpy==0.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.28
protobuf==4.24.4
pssh==2.3.4
psutil==5.9.0
psycopg2==2.9.2
psycopg2-binary==2.9.7
ptyprocess==0.7.0
py==1.10.0
py-cpuinfo==9.0.0
py3dns==3.2.1
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.1
pybullet==3.2.5
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
pycurl==7.44.1
pydot==1.4.2
pyflakes==2.4.0
pyglet==1.5.14
Pygments==2.11.2
PyGObject==3.42.1
pygraphviz==1.7
PyJWT==2.3.0
pykerberos==1.1.14
pylint==2.12.2
pyls-black==0.4.6
pyls-spyder==0.4.0
pynrrd==0.4.3
pynwb==2.2.0
pyparsing==2.4.7
PyPDF2==1.26.0
pypng==0.0.20
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyrsistent==0.18.1
pyserial==3.5
PySimpleSOAP==1.16.2
pysvn==1.9.15
pytest==6.2.5
pytest-pylint==0.18.0
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-debianbts==3.2.0
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-ldap==3.2.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.3.3
python-vagrant==0.5.15
pythran==0.10.0
pytz==2022.1
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.3.0
QDarkStyle==2.8.1
QtAwesome==1.1.1
qtconsole==5.2.2
QtPy==2.0.0
Recoll==1.31.6
recollchm==0.8.4.1+git
regex==2021.11.10
reportbug===11.4.1ubuntu1
reportlab==3.6.8
requests==2.25.1
requests-kerberos==0.12.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
roman==3.3
rpy2==3.4.5
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
s3transfer==0.3.7
sagemath-standard==9.5
scikit-build==0.17.6
scikit-image==0.21.0
scikit-learn==0.23.2
scipy==1.8.0
SCons==4.0.1
seaborn==0.11.2
SecretStorage==3.3.1
semver==3.0.1
Send2Trash==1.8.1b0
SimpleITK==2.3.0
simplejson==3.19.1
six==1.16.0
sklearn-pandas==2.0.3
snowballstemmer==2.2.0
sortedcontainers==2.1.0
Sphinx==4.3.2
spyder==4.2.1
spyder-kernels==1.10.2
SQLAlchemy==1.4.31
SQLAlchemy-i18n==1.1.0
SQLAlchemy-Utils==0.37.8
statsmodels==0.14.0
stevedore==3.5.0
swiglpk==4.65.1
sympy==1.9
systemd-python==234
tables==3.8.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
terminado==0.13.1
testpath==0.5.0
textdistance==4.2.2
texttable==1.6.4
thonny==3.3.14
threadpoolctl==3.1.0
three-merge==0.1.1
tifffile==2023.9.18
toml==0.10.2
tomli==1.2.2
torch==1.8.0a0+unknown
torchaudio==0.7.0a0
torchvision==0.8.0a0
tornado==6.1
tox==3.21.4
tqdm==4.57.0
traitlets==5.1.1
typed-ast==1.4.3
typing-extensions==3.10.0.2
tzlocal==2.1
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
ujson==5.1.0
unicodedata2==14.0.0
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.13.0+ds
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.4
wadllib==1.3.6
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==3.0.1
widgetsnbextension==2.0.0
wrapt==1.13.3
wurlitzer==3.0.2
xarray==2023.8.0
xdg==5
xkit==0.0.0
yarl==1.9.2
zipp==1.0.0
